EXP-138: Apply TryCallGuestFunction RAX propagation fix — 6 changes across 2 files
Root cause fix for EXP-026/EXP-137 '232 NULL returns':
- CallNativeEntry: int -> ulong (preserve 64-bit function pointers)
- ExecuteGuestThreadEntry: write context.Rax = nativeReturn after thunk
- ExecuteGuestContinuationEntry: same Rax write-back
- Entry path (num6): int -> ulong, -1 -> ulong.MaxValue
- NativeWorker RunGuestEntryStub: int -> ulong (dead code, type consistency)

MANDATORY regression gate: Dreaming Sarah → Arise → Yatzi (in order).
Do NOT skip regression gate — highest blast radius fix in project.

Build + runtime tests PENDING (no dotnet SDK in sandbox).
See exp-reports/EXP-138-final.md for maintainer instructions.

+++ work/sharpemuT24/src/SharpEmu.Core/Cpu/Native/DirectExecutionBackend.NativeWorker.cs
@@ -56,7 +56,7 @@ public sealed partial class DirectExecutionBackend
 	// Callers set the Active* thread-statics before emitting the stub and read the
 	// yield/forced-exit flags right after this returns, so the worker outcome is
 	// copied back into this thread's statics before returning.
-	private unsafe int RunGuestEntryStub(void* entryStub, ulong hostRspSlot)
+	private unsafe ulong RunGuestEntryStub(void* entryStub, ulong hostRspSlot)
 	{
 		var worker = RentNativeGuestExecutor();
 		if (worker is null)
@@ -82,7 +82,13 @@ public sealed partial class DirectExecutionBackend
 			_activeGuestThreadYieldRequested = yieldRequested;
 			_activeGuestThreadYieldReason = yieldReason;
 			_activeForcedGuestExit = forcedExit;
-			return nativeReturn;
+			// EXP-138: Cast int -> ulong for type consistency with CallNativeEntry.
+			// NOTE: The Windows native worker stub (line ~378) currently captures
+			// eax (32-bit) not rax (64-bit), so the upper 32 bits are already lost
+			// before this cast. The native stub fix (mov edx,eax -> mov rdx,rax)
+			// is a separate follow-up for Windows correctness; on Linux the native
+			// worker path is not used (RentNativeGuestExecutor returns null).
+			return (ulong)nativeReturn;
 		}
 		finally
 		{
